+++ pkg/models/memory.go
@@ -1,8 +1,8 @@
 package models
 
 type MemoryCurrent struct {
-	Units  string `json:"units"`
-	Memory uint64 `json:"memory"`
+	Units  string  `json:"units"`
+	Memory float64 `json:"memory"`
 }
 
 type MemoryHistory struct {
